% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parents.R
\name{create_parents}
\alias{create_parents}
\title{Create parents.}
\usage{
create_parents(n, len, homozygous = TRUE)
}
\arguments{
\item{n}{Number of parents to create.}

\item{len}{Vector of chromosome lengths in cM.}

\item{homozygous}{Should the parents be considered completely inbred?}
}
\value{
A nested list of lists where the first level refers
to the individuals and the second level to the different
chromosomes.
}
\description{
Create parents.
}
\examples{
create_parents(n = 2, L = c(100, 200), inbred = FALSE)

}
\author{
Dominik Mueller (\email{dominikmueller64@yahoo.de})
}
\seealso{
\code{\link{cross}}, \code{\link{sim_meiosis}}
}

