% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{cross}
\alias{cross}
\alias{dh}
\alias{gamete}
\alias{self}
\alias{sim_meiosis}
\title{Cross individuals, self individual, create doubled haploid.}
\usage{
cross(mother, father, xo_params)

self(parent, xo_params)

dh(parent, xo_params)

gamete(parent, xo_params)

sim_meiosis(parent, xo_params)
}
\arguments{
\item{mother}{An individual.}

\item{father}{Another individual.}

\item{xo_params}{A list with crossover simulation parameters as produced by
\code{\link{create_xo_params}}.}

\item{parent}{An individual.}

\item{parent}{An individual.}

\item{parent}{An individual.}

\item{parent}{An individual.}
}
\value{
A list of the progeny's chromosomes. For \code{dh}, this is reduced to a single
chromatid to save memory space.
}
\description{
Simulate the cross of two individuals.

Self an individual.

Derive a doubled haploid from an indivdiual.

Derive a gamete from an indivdiual.

Simulate meiosis.
}
\examples{
L <- c(100, 200)
xo_params <- create_xo_params(L)
cross(parents[[1]], parents[[2]], xo_params)
self(parents[[1]], xo_params)
dh(parents[[1]], xo_params)
gamete(parents[[1]], xo_params)
sim_meiosis(parents[[1]], xo_params)
}
\author{
Dominik Mueller (\email{dominikmueller64@yahoo.de})
}
\seealso{
\code{\link{create_xo_params}}, \code{\link{create_parents}}
}

