% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realized_IBD.R
\name{realized_IBD}
\alias{realized_F}
\alias{realized_IBD}
\alias{realized_IBD_matrix}
\title{Compute realized IBD coefficients.}
\usage{
realized_IBD(ind1, ind2)

realized_F(ind)

realized_IBD_matrix(pop, second_pop = NULL)
}
\arguments{
\item{ind1}{first individual}

\item{ind2}{second individual}

\item{ind}{individual}

\item{pop}{A population as a list of indivdiuals.}

\item{second_pop}{A second population.}
}
\description{
Compute realized IBD coefficients between two individuals.

Compute the realized inbreeding coefficient of an individual.

Compute realized IBD coefficients within one or between two
populations.
}
\details{
The realized IBD coefficient between two individuals is here defined as
the probability that two alleles randomly drawn each from one of the individuals
are IBD, i.e., refer to the same founder allele. This is analogous to the definintion of
the coefficient of coancestry in classical quantitative genetics, but the difference is
that while the coefficient of coancestry is calculated conditional on the pedigree and
hence can be regarded as an expected IBD coefficient, the realized IBD coefficient
takes into account mendelian sampling.
}
\examples{
L <- c(100, 200)
parents <- create_parents(2, L)
xo_params <- create_xo_params(L)
F1 <- cross(parents[[1]], parents[[2]], xo_params)
realized_IBD(self(F1, xo_params), self(F1, xo_params))
NULL
# After three generations of selfing.
realized_F(self(self(self(F1, xo_params), xo_params), xo_params))
NULL
pop <- replicate(4, self(F1, xo_params), simplify = FALSE)
realized_IBD_matrix(pop)
}
\author{
Dominik Mueller (\email{dominikmueller64@yahoo.de})
}

