% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{sim_founder}
\alias{sim_founder}
\title{Simualte founder.}
\usage{
sim_founder(n, m)
}
\arguments{
\item{n}{An integer. The number of founders.}

\item{m}{A integer vector. The number of loci on each chromosome.}
}
\value{
A list of matrices with biallelic loci, coded with 0 and 1.
}
\description{
Simulate founder genotypes.
}
\details{
This function is only for testing purposes.
}
\examples{
n <- 2L
m <- c(5L, 7L, 3L)
sim_founder(n ,m)

}
\author{
Dominik Mueller (\email{dominikmueller64@yahoo.de})
}

